A minor, and probably irrelevant correction for boundary interpolation
The coefficients for interpolation of phi(d1,d2) in the upper boundary of
d1 and d2 have been corrected in vdwxc:bcucof. But since these boundary
values may be never encountered, this correction is probably irrelevant.
Additionally, a couple of comments have been corrected, and the reference 
output of the graphite_vdw_df test has been updated.

modified:
  Src/SiestaXC/vdwxc.F90
  Tests/Reference/graphite_vdw_df.out
  Util/SiestaSubroutine/README