Optionally avoid mixing DM or H after scf-convergence
* If the option

   SCF.MixAfterConvergence F

is used, the program will exit the scf loop without a final mixing.

The DM coming out of the cycle will then be DM_out (regardless of whether
we are mixing H or the DM).

The hamiltonian coming out of the cycle is H_out (generated with
DM_out), but H_in ("Hold" in the code) is also kept if the option is
set, so it could be reset if needed in the future.

This feature will have an effect on the quality of the final E_KS energy
and of the forces when mixing the DM.

* The DM is saved during the scf cycle for restarting purposes (unless
overridden by the user). If there is no mix after convergence, the
final DM_out is saved.

(+ replaced the "siesta:" prefix by "scf:" in the relevant scf cycle output)

(+ changed the order of the monitor_forces_in_scf block, so that it is executed
before mixing).

(+ Dold is now saved in compute_dm, and Hold in setup_hamiltonian)
(+ Removed Dold code in final_H_f_stress )

added:
  Src/compute_max_diff.F
  Src/save_density_matrix.F
modified:
  Docs/Mixing.CHANGES
  Src/Makefile
  Src/compute_dm.F
  Src/final_H_f_stress.F
  Src/m_pulay.F90
  Src/mixer.F
  Src/scfconvergence_test.F
  Src/setup_hamiltonian.F
  Src/siesta_forces.F
  Src/siesta_options.F90
  Src/write_subs.F