Differences between tbtrans and tbtrans_rep
This fixes some differences reported between the
tbtrans and tbtrans_rep codes.
Essentially the difference was due to the integration points
which were not distributed equivalently in the two examples.
Now the contours are equivalent. There are still differences
but those may be "ironed" out by increasing the number of
integration points. So these are merely numerical artefacts.

Secondly, the tbtrans_rep output printed energies in eV but
the DOS was in 1/Ry, this is now fixed. Although it should not
matter because it is the relative values that are important.